requirements.txt
aiofiles==24.1.0
arxiv==2.2.0
beautifulsoup4==4.13.5
bert_score==0.3.13
doclayout_yolo==0.0.4
PyMuPDF
gradio
langchain==0.3.27
langchain_openai==0.3.33
matplotlib
numpy
openai==1.108.1
pandas==2.3.2
pdfplumber==0.11.7
Pillow==11.3.0
prettytable==3.16.0
pydantic==2.11.9
pytesseract==0.3.13
python-dotenv==1.1.1
PyYAML==6.0.2
rouge_score==0.1.2
scipy==1.16.2
seaborn==0.13.2
simpledorff==0.0.2
tiktoken==0.11.0
tqdm==4.67.1
lxml